+++ src/controller/src/controller/helpers.hpp
@@ -10,9 +10,9 @@
 
 namespace store
 {
-    class IStockStore;        // Forward declaration
-    class TransactionStore;   // Forward declaration
-    class PositionStore;      // Forward declaration
+    class IStockStore;         // Forward declaration
+    class ITransactionStore;   // Forward declaration
+    class PositionStore;       // Forward declaration
 }   // namespace store
 
 namespace drafts
@@ -30,16 +30,16 @@ namespace controller
     );
 
     std::vector<drafts::PositionDetailDraft> getOpenPositionDrafts(
-        AccountId                                  account,
-        const store::PositionStore&                positionStore,
-        const std::shared_ptr<store::IStockStore>& stockStore,
-        const store::TransactionStore&             transactionStore
+        AccountId                                        account,
+        const store::PositionStore&                      positionStore,
+        const std::shared_ptr<store::IStockStore>&       stockStore,
+        const std::shared_ptr<store::ITransactionStore>& transactionStore
     );
 
     std::vector<drafts::PositionDraft> getOpenPositionDrafts(
-        const store::PositionStore&                positionStore,
-        const std::shared_ptr<store::IStockStore>& stockStore,
-        const store::TransactionStore&             transactionStore
+        const store::PositionStore&                      positionStore,
+        const std::shared_ptr<store::IStockStore>&       stockStore,
+        const std::shared_ptr<store::ITransactionStore>& transactionStore
     );
 
 }   // namespace controller