+++ src/store/src/store/account/account_session.hpp
@@ -1,5 +1,5 @@
-#ifndef __STORE__INCLUDE__STORE__ACCOUNT__ACCOUNT_SESSION_HPP__
-#define __STORE__INCLUDE__STORE__ACCOUNT__ACCOUNT_SESSION_HPP__
+#ifndef __STORE__SRC__STORE__ACCOUNT__ACCOUNT_SESSION_HPP__
+#define __STORE__SRC__STORE__ACCOUNT__ACCOUNT_SESSION_HPP__
 
 #include <vector>
 
@@ -31,4 +31,4 @@ namespace store
     };
 }   // namespace store
 
-#endif   // __STORE__INCLUDE__STORE__ACCOUNT__ACCOUNT_SESSION_HPP__
+#endif   // __STORE__SRC__STORE__ACCOUNT__ACCOUNT_SESSION_HPP__